requirements.txt
flask
langchain
python-dotenv
langchain-google-genai
